+++ resources/migrations/20221003054324-add-uhour.up.sql
@@ -1,0 +1,1 @@
+ALTER TABLE users ADD COLUMN uhour VARCHAR(10);