+++ src/clj/l22/routes/home.clj
@@ -9,22 +9,8 @@
    [l22.password :refer [password password!]]
    [ring.util.response]))
 
-(def ^:private version "0.6.0")
-(def ^:private updated_at "2022-08-06 09:36:17")
-;; below only works in development, not in jar.
-;; (def ^:private version
-;;   (-> "project.clj" slurp read-string (nth 2)))
-
-;; cancel by 0.2.14
-;; the reason? cors?
-;; (defn home-page [{:keys [flash] :as request}]
-;;   (let [body (-> (client/get "https://w.hkim.jp/loc")
-;;                  :body
-;;                  (json/read-str :key-fn keyword))]
-;;     (layout/render request "home.html"
-;;                    {:flash flash
-;;                     :loc (:location body)
-;;                     :ts (:timestamp body)})))
+(def ^:private version "0.8.5")
+(def ^:private updated_at "2022-10-04 19:57:31")
 
 (defn home-page [{:keys [flash] :as request}]
   (layout/render request "home.html"
@@ -33,8 +19,6 @@
 (defn about-page [request]
   (layout/render request "about.html" {:version version
                                        :updated_at updated_at}))
-
-
 
 (defn home-routes []
   [""